+++ OpenMarket/OpenMarket/Model/ProductDetail.swift
@@ -15,9 +15,9 @@ struct ProductDetail: Codable {
     let description: String
     let thumbnail: String
     let currency: String
-    let price: Int
-    let bargainPrice: Int
-    let discountedPrice: Int
+    let price: Double
+    let bargainPrice: Double
+    let discountedPrice: Double
     let stock: Int
     let createdAt: String
     let issuedAt: String
